diffusion 2.0 float

+++ dsipts/models/Diffusion.py
@@ -310,10 +310,10 @@ class Diffusion(Base):
             #   params = sum([np.prod(p.size()) for p in model_parameters]) -> 13K
             true_log_var_clipped = _extract_into_tensor( self.posterior_log_variance_clipped, t, y_noised.shape )
             nonzero_mask = float((t != 0))  # no adding noise when t == 0
-            # variance range if it is learned (constant values, so out of the for cycle)
 
             if self.learn_var:
                 eps_pred, var_pred = sub_net(y_noised, y_past, emb_cat_past, emb_cat_fut, aux_emb_num_past, aux_emb_num_fut)
+                # variance range if it is learned (constant values, so out of the for cycle)
                 var_range_A = _extract_into_tensor(np.log(self.betas) , t, eps_pred.shape)
                 var_range_B = true_log_var_clipped
                 out_log_var = torch.exp(var_pred*var_range_A + (1-var_pred)*var_range_B)
@@ -574,7 +574,7 @@ class SubNet(nn.Module):
         eps_hat = y_noised - y_noised_hat
 
         # emb_eps_hat for further computations
-        emb_eps_hat = self.lin_eps_d_model(eps_hat) # -> [B, future_step, d_model]
+        emb_eps_hat = self.lin_eps_d_model(eps_hat.float()) # -> [B, future_step, d_model]
 
         # emb_eps_hat updated according to changes of CATEGORICAL information
         # needed info about both past and future
